add ID_TYPE to the id probers
Export the type of device from ata_id and scsi_id, also strip
leading and trailing whitespace and substitute consecutive
whitespace with a single underline char.

From: Hannes Reinecke <hare@suse.de>

+++ udev_utils.c
@@ -485,7 +485,7 @@ int execute_program(const char *command, const char *subsystem,
 
 				len += count;
 				if (len >= ressize-1) {
-					err("ressize %d too short", ressize);
+					err("ressize %ld too short", (long)ressize);
 					retval = -1;
 					break;
 				}
